+++ src/Products/PloneMeeting/browser/overrides.py
@@ -528,8 +528,9 @@ class PMRenderCategoryView(IDRenderCategoryView):
         default_template = self.cfg.get_default_item_template()
         default_template_uid = None
         if default_template and \
-           set(self.tool.get_orgs_for_user()).intersection(
-                default_template.getTemplateUsingGroups()):
+           (not default_template.getTemplateUsingGroups() or
+            set(self.tool.get_orgs_for_user()).intersection(
+                default_template.getTemplateUsingGroups())):
             default_template_uid = default_template.UID()
         return default_template_uid
 
@@ -711,8 +712,12 @@ class MeetingItemActionsPanelView(BaseActionsPanelView):
         if self.context.query_state() == 'validated':
             isPresentable = self.context.wfConditions().mayPresent()
 
+        # this volatile is invalidated when user/groups changed
+        date = get_cachekey_volatile(
+            'Products.PloneMeeting.ToolPloneMeeting._users_groups_value')
+
         # check also portal_url in case application is accessed thru different URI
-        return (repr(self.context), self.context.modified(), advicesIndexModified,
+        return (repr(self.context), self.context.modified(), advicesIndexModified, date,
                 sent_to,
                 isRealManager, isManager, isEditorUser,
                 userAbleToCorrectItemWaitingAdvices, isPowerObserverHiddenHistory,
@@ -794,7 +799,8 @@ class MeetingActionsPanelView(BaseActionsPanelView):
                                    'renderTransitions',
                                    'renderOwnDelete',
                                    'renderDeleteWholeMeeting',
-                                   'renderActions', ]
+                                   'renderActions',
+                                   'renderHistory']
 
     def __call___cachekey(method,
                           self,